talents in config file
