+++ wow_simulator_web/simulator_interface/templates/components_new/talents.html
@@ -1,11 +1,10 @@
 <h3><i>Talents</i></h3>
 <div class="rounded border" style="padding-left:50px; padding-right:50px; padding-top:20px; padding-bottom:20px; border-color: darkgrey !important;">
-    {% for field in form|filteritems:'talents_' %}
+    {% for field in form|filteritems:'talents-' %}
         {% if forloop.first %}<div class="row">{% endif %}
         <div class="col-sm-6" style="text-align:left;">
             {{ field }} <label for="{{ field.id }}" class="control-label"> {{ field.label }} </label>
-
-<!--            <input type="hidden" value="{{ talent.0 }}" name="talent_names" />-->
+            <input type="hidden" value="{{ field.id }}" name="talent_names" />
         </div>
         {% if forloop.counter|divisibleby:2 %}</div><br><div class="row">{% endif %}
         {% if forloop.last %}</div>{% endif %}
